transform multipart
